miniciod: add DataStream functions to try and get output working.
No joy, and apparently if I send any output the node wedges when it tries to
terminate, making me think maybe I'm stalled in the middle of communications
transaction.  Need to understand the Cioman stuff better if I'm going to get
this to work.  The actual output is commented out so it won't cause others
to wedge.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>