multipipe: Plan 9 native multipipe device
This is a basic port from the Inferno multipipe device
driver with some fundamental improvements.  Its implements
the FIFO style mpipe as opposed to any of the other types, but
we'll implement deterministic hash pipes and others in the future.

This is a work-in-progress, but I wanted to checkpoint initial revision
to source control.  It does not currently build cleanly as I have to
implement read and write hooks.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>