execfs: first cut at Plan 9 native exec extension
This is the begining of a file system overlay to add clone
semantics to proc(3).  It acts as an overlay, only providing
the clone file with the intent that it can mostly use the underlying
proc(3) for control and status interfaces.

We need to add some more paranoid error checking code, support for
arguments, and perhaps most importantly add support to add stdio
pipes (which get bound over the pid at the appropriate time).

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>