mpipefs: added broadcast mode to pipes
This patch adds broadcast mode to pipes (which is specified at
attach time).  Writes to a broadcast pipe are replicated to all
readers.  Writes with no readers go to the bitbucket (which may
not be the right or expected functionality).  Multi-packet writes
aren't supported as such, but don't make much sense anyways since
all writes go to all places.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>