infrastructure: debug mode
add a debug environment variable as a standard item.  If its on,
you won't get output in the standard way and you'll need to use
mydebug (a modified version of jtagdebug) and trigger the redirect
console commands (and subsequent write_con) sequences.  This will
help with getting prints which don't go out via kprint and will also
get rid of problems with race conditions in the log aggregation code.

This code currently is not completely tested due to issues with the
kernel crashing early in boot up.  But .... it should work :)

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>