Added &0xffff to NETQID to keep rest of bits clean since NETID only uses the lower 16 bits.  This is possibly not necessary for proper functioning, but as we deal with the qid uses operating across the synthetic file systems I will want to know exactly what portions are the NETID, PID, or other delineators.
Changed MAGIC to use hex definition instead of decimal.  Note: it was
defined as a hex number in gangfs, and makes it easier to know how
many bits are actually being used without converting the number.

Added the bottom 11 bits of mpipefs's PID to the NETQID.  It currently
uses getpid() directly and does not cache the pid.  This might be a
little expensive, and we might want to consider using a cached version
of the PID instead.