cleaned up some of the debugging IO
segfault caused by under allocating resources due to using "avail >
remaining" instead of >=.  This would set up an empty session and the
next loop triggers an assert where nprocs available = 0.  This was
fixed by adding the = in the test, and also simply continuing to the
next child if this happens (there are still other children whom can
service the task)...

The first time in resgang, the resources where not computed so I added
a call to refreshstatus.

moved a second check for m==nil in statuswrite to where it does some
good.