fix a race condition in cloneproc where an execfs's ctl file may not be fully stable before it is opened and read.  This is moved into a tight spin-retry-loop.
add some notes in resgang -- if a session fails, then the whole node
is marked as broken and resgang returns an error.  It should fail-safe
if there are sufficient resources and continue.  Fixing the cloneproc
race condition should solve the immediate problem, but there needs to
be a forward thinking solution which accommodates session and node
failures.